Changed database location from 'config directory' to 'data directory'
This should be teh correct location and also makes the division between
config/data clear. Cleaning config directory will not remove the
vocabulary and cleaning the data directory will not remove the
configuration.

The database will not be moved automatically. Since there is no
officially released version, this should not be a problem. Anyone using
the current version has to move the database by himself.