New database schema
This schema allows for adding multiple vocabulary with the same 'word'
as well as giving each vocabulary a language.

Language interface is not yet implemented but should be possible in the
future.

https://github.com/Top-Ranger/harbour-vocabulary/issues/10

PLEASE NOTE: This schema has the same version number as the prior one
(since the prior was not part of an official release). If you have used
the prior version 3 one (e.g. a self compiled unstable version), export
your vocabulary, delete the database and import them again to
avoid problems.