Word priority visible in vocabulary list
Just one more idea.
It would be useful to have the overview of the priority score of each word in the "Show all vocabulary" view.
After this change, the priority should be visible as a bar at the bottom of the row. (It is untested, I got the inspiration and code concept from gPodder).